Fixed a couple of warnings and grammar mistakes

+++ src/main/java/com/b2/payload/response/AddressResponse.java
@@ -17,5 +17,5 @@ public class AddressResponse {
     private String city;
 
     @NotNull
-    private int probability;
+    private Float probability;
 }
